requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==0.16
Jinja2==2.7
MarkupSafe==0.18
MySQL-python==1.2.4
SQLAlchemy==0.8.1
Tempita==0.5.1
Werkzeug==0.9.1
decorator==3.4.0
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
itsdangerous==0.21
passlib==1.6.1
requests==1.2.0
wsgiref==0.1.2
